requirements.txt
certifi==2020.4.5.2
chardet==3.0.4
coverage==5.1
idna==2.9
Pillow==7.1.2
requests==2.24.0
six==1.15.0
urllib3==1.25.9
vk-api==11.8.0

pony~=0.7.13